style: improve overall look & feel of app
also add some UX techniques to improve it

+++ src/components/delete-todo.tsx
@@ -1,14 +1,22 @@
 "use client";
 import { deleteTodoAction } from "@/server/actions/delete-todo.action";
+import { Trash2 } from "lucide-react";
 
 export function DeleteTodo({ id }: Readonly<{ id: number }>) {
   const handleDelete = async () => {
-    const result = await deleteTodoAction(id);
-    console.log(result);
+    if (confirm("Are you sure you want to delete this task?")) {
+      await deleteTodoAction(id);
+    }
   };
 
   return (
-    <button onClick={() => handleDelete()} type="button">
+    <button 
+      onClick={handleDelete} 
+      type="button"
+      className="text-red-500 hover:text-red-700 font-medium flex items-center text-sm focus:outline-none"
+      aria-label="Delete task"
+    >
+      <Trash2 className="h-4 w-4 mr-1" />
       <span>Delete</span>
     </button>
   );
